Made "name" argument mandatory in OpenRegistrationCommand.
Name is no longer an optional argument for "!event-registration-open".
This is because when name isn't provided, Event.findCurrentEvent() is
utilised, which looks for events which are either open for registration
or checkin, but an event which is about to be opened for registration
is neither open for registration or checkin, so Event.findCurrentEvent()
will never return an event which isn't opened for registration yet.